Move the fhirendpointquerier to endpointmanager/pkg.
