+++ endpointmanager/cmd/endptquerier/main.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/config"
-	endptQuerier "github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointmanager/fhirendpointquerier"
+	endptQuerier "github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/fhirendpointquerier"
 	"github.com/onc-healthit/lantern-back-end/endpoints/fetcher"
 
 	log "github.com/sirupsen/logrus"
